+++ src/main/java/ru/yandex/practicum/filmorate/mapper/MpaRowMapper.java
@@ -12,7 +12,7 @@ public class MpaRowMapper implements RowMapper<Mpa> {
     @Override
     public Mpa mapRow(ResultSet rs, int rowNum) throws SQLException {
         return Mpa.builder()
-                .id(rs.getInt("id"))
+                .id(rs.getLong("id"))
                 .name(rs.getString("mpa"))
                 .build();
     }
